+++ NEWS.md
@@ -6,6 +6,7 @@
 - used native HTTP client with OpenSSL 3.0 or later, removed libcurl dependency
 - added the "-ignore-crl" option to disable fetching and verifying
   CRL Distribution Points
+- changed error output to stderr instead of stdout
 - improved testing
 
 ### 2.8 (2024.03.03)